requirements.txt
PyQt5==5.15.9
   python-docx==0.8.11
   openai==0.28.0
   google-generativeai==0.1.0
   chardet==5.1.0
